refactor: extracted types into its own file and updated dependent files
