+++ src/lib/types.ts
@@ -1,0 +1,20 @@
+export type categories =
+  | "café"
+  | "bar"
+  | "brewery"
+  | "hike"
+  | "restaurant"
+  | "shop"
+  | "park"
+  | "other";
+
+export interface Location {
+  id: string;
+  name: string;
+  address: string;
+  categories: categories[];
+  lat: number;
+  lng: number;
+  website?: string;
+  photo?: string;
+}
